[fix] Builders check primitives are initialized (#36)
## Before this PR

Conjure POJOs containing primitives (e.g. int, boolean) would be coerced from JSON in which the fields were actually missing.  This was not compliant with the wire spec and could lead to nasty bugs, where a field was unintentionally initialized to `0` or `false`. For example, consider this type:

```yaml
UpdateSalaryRequest:
  salary: integer
  shares: integer
```

The following request would result in dfox's salary being set to 0, because the the 'salary' field was missing and is initialized to 0!!
```
curl -XPOST https://some-api/update-salary/dfox --data '{"shares":20}'
```


## After this PR

Trying to deserialize a type with booleans, integers or doubles from `{}` will now fail.

## Concerns

People may have been relying on this behaviour, so there's a chance this might break people's servers... In theory, we could put this behaviour behind a feature flag, and then phase it in slowly?? EDIT not gonna do this as it's not actually a wire-break and people's servers can be trivially fixed by updating their conjure defs:

```diff
 UpdateSalaryRequest:
-  salary: integer
+  salary: optional<integer>
   shares: integer
```


_Future work: unify this `validatePrimitivesHaveBeenInitialized` method with the other `validateFields` method_

fixes #14