Retrofit2 clients should also set hr-path-template
For tracing purposes (and soon, also concurrency limiting) the Feign
contract sets the hr-path-template header to the path template for
tracing purposes. An OkHttp interceptor removes this header, and
uses it to start a span.

Feign does this, and Retrofit does not. Let's make our retrofit
interfaces contain this header, too!

I've traced this through with a debugger and verified that it does the
right thing.