[fix] optional<binary> equivalent to binary for retrofit + Optional<StreamingOutput> for jersey (#201)
Fix #182 

For jersey and retrofit, the current behavior of `binary` enables an implicit tri-state with different status code:
- non-empty stream: 200
- empty stream: 200
- null: 204 + null stream on retrofit client

The way forward to achieve a standard behavior with all clients and server is to consider that service that take advantage of this behavior should declare a return type of `optional<binary>` instead. The jersey resource implementation will need to stay the same and keep encoding empty as null.

In this PR, we are making the return type `optional<binary>` being equivalent to `binary` for the retrofit generator and change the jersey generator to return `Optional<StreamingOutput>`.