[improvement] Simpler alias valueOf implementation
We no longer duplicate code between `of` and `valueOf`. `valueOf` simply
parses a string and passes the result to `of`.
Alias.of(null) results in a safe-logging npe rather than using
`Objects.requireNonNull`. Slightly simplified constructors to use
the result of the null check.
Add a `valueOf` function to alias of external type imports which both
declare a base type allowed to have a valueOf, and have a
`public static valueOf` function on the imported type (only applies
to types known by the generator).