Fix deserialization for aliases of external type imports
JsonCreator factory methods are hanled as 'native' deserializers,
bypassing all ObjectMapper configuration and deserializers.
https://github.com/FasterXML/jackson-databind/issues/2318

In cases when both a valueOf and of method exist for an
external type import, we do not apply the `JsonCreator`
annotation, allowing jackson to choose the most accurate
deserializer function (either `of` or `valueOf` depending
on the input node type).