[improvement][undertow] Add EndpointHandlerWrapper for filter-like features (#406)
## Changes in this PR

This PR enables features that are similar to the concept of Jersey filters. Undertow is meant to be used by chaining `HttpHandler`s. The first `HttpHandler` applies logic and then pass it to the next handler. With this structure, handlers always have control over the control flow and can apply logic, before and after processing by the next handlers. This is similar to the desired behavior of jersey Filters. 

After that PR, it is possible add arbitrary behaviours pre and post processing of the underlying resource by leveraging a concept similar to Undertow's `HandlerWrapper`: the `EndpointHandlerWrapper`. It wraps the handler of an endpoint, taking potentially advantage of all informations contained in an `Endpoint`.

Below an example of an EndpointHandlerWrapper with pre and post processing.

```java
EndpointHandlerWrapper myHandlerWrapper = 
    endpoint -> 
      if (shouldWrap) {
        Optional.of(exchange -> {
          preProcessingLogic(endpoint, exchange);
          endpoint.handler().handleRequest(exchange);
          postLogic(endpoint, exchange)
          exchange.addCompletionListener(postCompletionLogic(endpoint))})
      } else {
        Optional.empty()
      }
```

- added `EndpointHandlerWrapper` as a functional interface of the form `HttpHandler wrap(Endpoint endpoint)`. **Warning** `EndpointHandlerWrapper`s are meant to be used only by library writers and webserver maintainers.
- added `addEndpointHandlerWrapperBeforeBlocking(EndpointHandlerWrapper wrapper)` to the `ConjureHandler` class to give the possibility of adding **non-blocking** `EndpointHandlerWrapper` to all registered `UndertowService`s.

The utilities to easily wrap an `UndertowService` with an `EndpointHandlerWrapper` have been purposefully not included as this is a behavior that we do not want to encourage.
One can still use the following snipper to achieve this behavior.

```java
    public static UndertowService map(UndertowService service, EndpointHandlerWrapper wrapper) {
        return runtime ->
                ImmutableList.copyOf(
                        service.endpoints(runtime).stream()
                                .map(endpoint -> Endpoints.map(endpoint, wrapper))
                                .collect(ImmutableList.toImmutableList()));
    }
```

## Possible downsides?

None