conjure-undertow drains request bytes prior to writing responses
Note that we already effectively provide this behavior when the
response is sufficiently small (fits into some combination of the
web server and system buffers) where the request is drained after
the response has been sent. This change allows us to provide the
same behavior in the presence of large responses.