Support alias of wrapper null coercion
This is similar to an earlier change, but applies to aliases.
Note that this does allow `{}` to coerce into an empty value, however
it also allows null to work the same between aliases and dealiased
collections. It's more important to correctly support required
parts of the spec than to guard against other edge cases.

Also note that the tests exect null to result in a failure, however
null may be coerced into an empty collection.