Support exclusion of empty collections using `experimentalExcludeEmptyCollections`
This matches empty optional exclusion, including discovery
and remediation of a bug:

Aliases may be excluded, however they were not handled correctly
in the other direction, where null values were handled correctly.
Now we initialize builder instances with a reference to an empty
instance of alias types. This includes a new `public static T empty()`
method on aliases which can handle defaults.

The current implementation only checks one level of alias indirection,
so alias-of-alias-of-optional/collection still won't do quite what we
want, but it's a start.