Simplify SafeLong implementation
We use a simple final class rather than leveraging immutables
unnecessarily.

It's possible that an unknown piece of code has subclassed the
`abstract` SafeLong type, this is unlikely based on an internal
search with no results, and immutables being a clear indicator
that arbitrary subtypes are not expected.

Previously the ImmutableSafeLong type was public, there are three
known direct references which I am in the process of removing.
This introduces some risk, but the type is internal API and
not meant to be consumed directly. No uses appear to be in
core libraries, so removal should not be risky.