Jackson StreamConstraintExceptions parsing requests produce 422 status
Previously these resulted in 500 response codes, which aren't the
most accurate, and worse often result in retries which consume
large buffers. I considered using 413 "request entity too large"
here, but that's not entirely accurate as the entity itself
may be a reasonable size. We're refusing to parse an entity because
a value is much larger than we expect, which is a bit more more
nuanced than http response status codes can (or should) convey.