support deprecated-for-removal on dialogue code generation
For endpoints containing the "deprecated" field, we generate a
`@java.lang.Deprecated` annotation on the method. With this change, if
the endpoint additionally contains a tag named "deprecated-for-removal",
the generated annotation changes to `@Deprecated(forRemoval = true)`.

Note that this change only affects dialogue service code generation, and
does not alter the behavior for Undertow or Jersey code generation.