conjure-undertow streamed binary response bodies protect against early closure
There's a common pattern where folks use a try-with-resources block
on the response stream. While it's generally good practice to close our
resources, the webserver is responsible for this one, and closing it
in a finally block can prevent us from relaying errors to clients!
By blocking close calls to the delegate, we can allow folks to close
all the things, while providing accurate failure information to
clients :-]